feat: crear componente para actualizar cotización
